requirements.txt
streamlit>=1.37
numpy>=1.24
pandas>=2.0
librosa>=0.10
mutagen>=1.47
pyloudnorm>=0.1.1
faiss-cpu>=1.7
umap-learn>=0.5.6
scikit-learn>=1.3
sqlite-utils>=3.36
pyyaml>=6.0
tqdm>=4.66
lightgbm>=4.3
joblib>=1.3
